changes directory structure for cli
